% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suntimes.R
\name{suntimes_multiple}
\alias{suntimes_multiple}
\title{Get solar event times for multiple dates.}
\usage{
suntimes_multiple(lat, lon, dates = NULL, timezone = NULL)
}
\arguments{
\item{lat}{latitude in decimal degrees. Required.}

\item{lon}{longitude in decimal degrees. Required.}

\item{dates}{dates as vector string in YYYY-MM-DD format. If not present,
date defaults to current date. Optional.}

\item{timezone}{a character string containing a timezone to convert to.
R must recognise the name contained in the string as a time zone on your
system. Use OlsonNames() for valid timezones. If not present,
UTC is returned. Optional.}
}
\description{
Returns times for sunrise, sunset, and various other solar events for the
chosen location, dates, and timezone.
}
\examples{
suntimes_multiple(50.065471, -5.714856, c("2023-08-31", "2023-09-01"))
suntimes_multiple(50.065471, -5.714856, c("2023-08-31", "2023-09-01"), "Europe/London")
}
